refactor(site): add tabbed layout for single group page
